# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Envelope:
    properties:
      code:
        type: integer
      data: { }
      message:
        type: string
    type: object
  response.PageData:
    properties:
      items: { }
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  screening.ColumnView:
    properties:
      field:
        example: pe
        type: string
      label:
        example: 市盈率
        type: string
      unit:
        example: '%'
        type: string
    type: object
...
  screening.DeleteResultView:
    properties:
      deleted:
        example: true
        type: boolean
    type: object
  screening.FieldDictView:
    properties:
      fields:
        items:
          $ref: '#/definitions/screening.FieldView'
        type: array
    type: object
  screening.FieldView:
    properties:
      kind:
        description: |-
          Kind 决定该字段能接受哪些比较符：numeric 有全序，九个比较符全可用；
          categorical 无序，只剩相等与集合运算。Operators 已按此过滤，
          前端不必自己再推一遍。
        enum:
          - numeric
          - categorical
        example: numeric
        type: string
      label:
        example: 净资产收益率
        type: string
      name:
        description: Name 就是提交 criterionRequest.field 时要填的串。
        example: roe
        type: string
      operators:
        items:
          $ref: '#/definitions/screening.OperatorView'
        type: array
      unit:
        description: Unit 是量纲提示（% / 万元），供输入框渲染后缀；无量纲字段不下发该键。
        example: '%'
        type: string
    type: object
...
  screening.ResultSetView:
    properties:
      asOf:
        description: |-
          AsOf 必须回给前端：筛选是对某一天横截面的查询，盘后与次日早盘
          执行同一个模板结果不同是正常的，但用户看不到日期就只会觉得系统不稳定。
        example: "2024-05-31"
        type: string
      columns:
        items:
          $ref: '#/definitions/screening.ColumnView'
        type: array
      items:
        items:
          $ref: '#/definitions/screening.ResultRowView'
        type: array
      total:
        description: Total 是命中总数，可能大于 len(items)——超出条数上限的部分被截断。
        example: 128
        type: integer
      truncated:
        description: Truncated 为真表示 total 超过了本次的条数上限，页面应提示用户收紧条件。
        type: boolean
    type: object
  screening.RunRequest:
    properties:
      criteria:
        items:
          $ref: '#/definitions/screening.CriterionRequest'
        type: array
      limit:
        description: Limit 留空按 50 条，上限 500；超出部分由值对象直接收敛，不报错。
        example: 50
        type: integer
      sortDirection:
        enum:
          - asc
          - desc
        example: desc
        type: string
      sortField:
        example: total_mv
        type: string
    required:
      - criteria
    type: object
  screening.TemplateRequest:
    properties:
      criteria:
        description: |-
          Criteria 整份替换，没有「增量加一条」的形状：编辑表单一次提交整份列表，
          增量语义会让「用户删掉了一条」无法表达。最多 20 条。
        items:
          $ref: '#/definitions/screening.CriterionRequest'
        type: array
      description:
        type: string
      isPublic:
        description: IsPublic 置真即上架「选股策略广场」：任何登录用户都能查看并执行，但改不动。
        type: boolean
      limit:
        example: 50
        type: integer
      name:
        example: 低估值蓝筹
        type: string
      sortDirection:
        enum:
          - asc
          - desc
        example: desc
        type: string
      sortField:
        example: total_mv
        type: string
    type: object
  screening.TemplateView:
    properties:
      createdAt:
        type: string
      criteria:
        items:
          $ref: '#/definitions/screening.CriterionView'
        type: array
      description:
        type: string
      id:
        type: integer
      isPublic:
        type: boolean
      limit:
        example: 50
        type: integer
      name:
        type: string
      sortDirection:
        enum:
          - asc
          - desc
        example: desc
        type: string
      sortField:
        description: |-
          SortField / SortDirection 恒为已展开的有效值：模板未指定排序时，
          这里回的是兜底的 total_mv desc，而不是空串。
        example: total_mv
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: { }
  description: |-
    多智能体股票分析平台 HTTP 接口。

    所有响应共用同一个信封 {"code":0,"message":"ok","data":{...}}，
    code 为 0 表示成功，非 0 时 data 恒为 null，具体取值见 response 包。

    健康检查 GET /healthz 不在 /api/v1 之下，因此未收录于本文档。

    免责声明：本平台仅用于学习与研究，不构成投资建议。分析结论由大模型生成，
    可能包含事实性错误。过往表现不代表未来收益，投资有风险，可能损失本金。
  title: Trading Agents API
  version: "1.0"
paths:
...
  /screening/fields:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.FieldDictView'
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 查询可筛选字段字典
      tags:
        - 选股
  /screening/run:
    post:
      consumes:
        - application/json
      parameters:
        - description: 筛选条件、排序与条数
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/screening.RunRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.ResultSetView'
                type: object
        "400":
          description: 条件为空、超过 20 条、重复，或字段/比较符/排序不合法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "503":
          description: 行情或财务存储未配置，无法执行筛选
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 执行临时筛选
      tags:
        - 选股
  /screening/templates:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/screening.TemplateView'
                    type: array
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 列出我的选股模板
      tags:
        - 选股
    post:
      consumes:
        - application/json
      parameters:
        - description: 模板名、筛选条件、排序与条数
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/screening.TemplateRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.TemplateView'
                type: object
        "400":
          description: 模板名为空、条件为空，或字段/比较符/排序不合法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 已存在同名模板，或提交中含重复的筛选条件
          schema:
            $ref: '#/definitions/response.Envelope'
        "429":
          description: 筛选条件超过 20 条上限
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 新建选股模板
      tags:
        - 选股
  /screening/templates/{id}:
    delete:
      parameters:
        - description: 模板 ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.DeleteResultView'
                type: object
        "400":
          description: 模板 ID 非正整数
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 模板不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 删除选股模板
      tags:
        - 选股
    get:
      parameters:
        - description: 模板 ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.TemplateView'
                type: object
        "400":
          description: 模板 ID 非正整数
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 模板不存在，或既非本人所有也未公开
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 查询选股模板详情
      tags:
        - 选股
    put:
      consumes:
        - application/json
      parameters:
        - description: 模板 ID
          in: path
          name: id
          required: true
          type: integer
        - description: 待更新的字段，留空表示不改
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/screening.TemplateRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.TemplateView'
                type: object
        "400":
          description: 模板 ID 非正整数，或字段/比较符/排序不合法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 模板不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 重名、条件重复，或模板已被并发修改
          schema:
            $ref: '#/definitions/response.Envelope'
        "429":
          description: 筛选条件超过 20 条上限
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 更新选股模板
      tags:
        - 选股
  /screening/templates/{id}/run:
    post:
      parameters:
        - description: 模板 ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/screening.ResultSetView'
                type: object
        "400":
          description: 模板 ID 非正整数，或模板内含无法识别的字段
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 模板不存在，或既非本人所有也未公开
          schema:
            $ref: '#/definitions/response.Envelope'
        "503":
          description: 行情或财务存储未配置，无法执行筛选
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 执行已保存的选股模板
      tags:
        - 选股
  /screening/templates/public:
    get:
      parameters:
        - description: 页码，从 1 开始，默认 1
          in: query
          name: page
          type: integer
        - description: 每页条数，默认与上限由分页值对象收敛
          in: query
          name: pageSize
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/response.PageData'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/screening.TemplateView'
                            type: array
                        type: object
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 列出公开选股模板
      tags:
        - 选股
...
securityDefinitions:
  BearerAuth:
    description: 填入 "Bearer {accessToken}"，令牌由 POST /auth/login 获取。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...